pyproject.toml
[tool.poetry]
name = "streamlit-webrtc-hugging-face-transformer-example"
version = "0.1.0"
description = ""
authors = ["Yuichiro Tachibana (Tsuchiya) <t.yic.yt@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "streamlit_webrtc_hugging_face_transformer_example"}]

[tool.poetry.dependencies]
python = "^3.9,<3.11,!=3.9.7"
streamlit = "^1.18.0"
streamlit-webrtc = "^0.44.2"
transformers = "^4.26.1"
torch = "^1.13.1"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py38']
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | streamlit_webrtc/frontend
)
'''

[tool.isort]
# https://pycqa.github.io/isort/docs/configuration/black_compatibility/
profile = "black"
